Add validation for user when delete a blog (ex4.21)

+++ controllers/blogs.js
@@ -61,6 +61,15 @@ blogsRouter
     const { id } = request.params;
 
     try {
+      const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+      if (!decodedToken.id) return response.status(401).json({ error: "token invalid" });
+
+      const user = await User.findById(decodedToken.id);
+      const blog = await Blog.findById(id);
+
+      if (user._id.toString() !== blog.user.toString()) return response.status(401).json({ error: "user invalid" });
+
       const deletedBlog = await Blog.findByIdAndDelete(id);
       response.json(deletedBlog);
     } catch (error) {
